requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
googletrans==4.0.0rc1
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
loadenv==0.1.1
packaging==24.0
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pypdfium2==4.29.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.34.0
rfc3986==1.5.0
rpds-py==0.18.0
sniffio==1.3.1
sqlparse==0.4.4
typing_extensions==4.11.0
uritemplate==4.1.1
